move test functions to their own file
